apps/queue/auth/requirements/core.txt
PyJWT
fastapi
sqlalchemy
psycopg2-binary
uvicorn[standard]
